fix: 메인 페이지 경로 변경
